Move model pdata interfaces to pdata, expose them publicly
