--- internal/model/encoder.go
@@ -1,36 +1,0 @@
-// Copyright The OpenTelemetry Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//       http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package model
-
-// MetricsEncoder is an interface to encode protocol-specific data model into bytes.
-type MetricsEncoder interface {
-	// EncodeMetrics encodes protocol-specific data model into bytes.
-	// If the error is not nil, the returned bytes slice cannot be used.
-	EncodeMetrics(model interface{}) ([]byte, error)
-}
-
-// TracesEncoder is an interface to encode protocol-specific data model into bytes.
-type TracesEncoder interface {
-	// EncodeTraces encodes protocol-specific data model into bytes.
-	// If the error is not nil, the returned bytes slice cannot be used.
-	EncodeTraces(model interface{}) ([]byte, error)
-}
-
-// LogsEncoder is an interface to encode protocol-specific data model into bytes.
-type LogsEncoder interface {
-	// EncodeLogs encodes protocol-specific data model into bytes.
-	// If the error is not nil, the returned bytes slice cannot be used.
-	EncodeLogs(model interface{}) ([]byte, error)
-}
