+++ internal/otlp/json_encoder.go
@@ -19,10 +19,10 @@ import (
 
 	"github.com/gogo/protobuf/jsonpb"
 
+	"go.opentelemetry.io/collector/consumer/pdata"
 	otlpcollectorlogs "go.opentelemetry.io/collector/internal/data/protogen/collector/logs/v1"
 	otlpcollectormetrics "go.opentelemetry.io/collector/internal/data/protogen/collector/metrics/v1"
 	otlpcollectortrace "go.opentelemetry.io/collector/internal/data/protogen/collector/trace/v1"
-	"go.opentelemetry.io/collector/internal/model"
 )
 
 type jsonEncoder struct {
@@ -36,7 +36,7 @@ func newJSONEncoder() *jsonEncoder {
 func (e *jsonEncoder) EncodeLogs(modelData interface{}) ([]byte, error) {
 	ld, ok := modelData.(*otlpcollectorlogs.ExportLogsServiceRequest)
 	if !ok {
-		return nil, model.NewErrIncompatibleType(&otlpcollectorlogs.ExportLogsServiceRequest{}, modelData)
+		return nil, pdata.NewErrIncompatibleType(&otlpcollectorlogs.ExportLogsServiceRequest{}, modelData)
 	}
 	buf := bytes.Buffer{}
 	err := e.delegate.Marshal(&buf, ld)
@@ -46,7 +46,7 @@ func (e *jsonEncoder) EncodeLogs(modelData interface{}) ([]byte, error) {
 func (e *jsonEncoder) EncodeMetrics(modelData interface{}) ([]byte, error) {
 	md, ok := modelData.(*otlpcollectormetrics.ExportMetricsServiceRequest)
 	if !ok {
-		return nil, model.NewErrIncompatibleType(&otlpcollectormetrics.ExportMetricsServiceRequest{}, modelData)
+		return nil, pdata.NewErrIncompatibleType(&otlpcollectormetrics.ExportMetricsServiceRequest{}, modelData)
 	}
 	buf := bytes.Buffer{}
 	err := e.delegate.Marshal(&buf, md)
@@ -56,7 +56,7 @@ func (e *jsonEncoder) EncodeMetrics(modelData interface{}) ([]byte, error) {
 func (e *jsonEncoder) EncodeTraces(modelData interface{}) ([]byte, error) {
 	td, ok := modelData.(*otlpcollectortrace.ExportTraceServiceRequest)
 	if !ok {
-		return nil, model.NewErrIncompatibleType(&otlpcollectortrace.ExportTraceServiceRequest{}, modelData)
+		return nil, pdata.NewErrIncompatibleType(&otlpcollectortrace.ExportTraceServiceRequest{}, modelData)
 	}
 	buf := bytes.Buffer{}
 	err := e.delegate.Marshal(&buf, td)
