+++ internal/otlp/to_translator.go
@@ -20,7 +20,6 @@ import (
 	otlpcollectorlogs "go.opentelemetry.io/collector/internal/data/protogen/collector/logs/v1"
 	otlpcollectormetrics "go.opentelemetry.io/collector/internal/data/protogen/collector/metrics/v1"
 	otlpcollectortrace "go.opentelemetry.io/collector/internal/data/protogen/collector/trace/v1"
-	"go.opentelemetry.io/collector/internal/model"
 )
 
 type toTranslator struct{}
@@ -32,7 +31,7 @@ func newToTranslator() *toTranslator {
 func (d *toTranslator) ToLogs(modelData interface{}) (pdata.Logs, error) {
 	ld, ok := modelData.(*otlpcollectorlogs.ExportLogsServiceRequest)
 	if !ok {
-		return pdata.Logs{}, model.NewErrIncompatibleType(&otlpcollectorlogs.ExportLogsServiceRequest{}, modelData)
+		return pdata.Logs{}, pdata.NewErrIncompatibleType(&otlpcollectorlogs.ExportLogsServiceRequest{}, modelData)
 	}
 	return pdata.LogsFromInternalRep(internal.LogsFromOtlp(ld)), nil
 }
@@ -40,7 +39,7 @@ func (d *toTranslator) ToLogs(modelData interface{}) (pdata.Logs, error) {
 func (d *toTranslator) ToMetrics(modelData interface{}) (pdata.Metrics, error) {
 	ld, ok := modelData.(*otlpcollectormetrics.ExportMetricsServiceRequest)
 	if !ok {
-		return pdata.Metrics{}, model.NewErrIncompatibleType(&otlpcollectormetrics.ExportMetricsServiceRequest{}, modelData)
+		return pdata.Metrics{}, pdata.NewErrIncompatibleType(&otlpcollectormetrics.ExportMetricsServiceRequest{}, modelData)
 	}
 	return pdata.MetricsFromInternalRep(internal.MetricsFromOtlp(ld)), nil
 }
@@ -48,7 +47,7 @@ func (d *toTranslator) ToMetrics(modelData interface{}) (pdata.Metrics, error) {
 func (d *toTranslator) ToTraces(modelData interface{}) (pdata.Traces, error) {
 	td, ok := modelData.(*otlpcollectortrace.ExportTraceServiceRequest)
 	if !ok {
-		return pdata.Traces{}, model.NewErrIncompatibleType(&otlpcollectortrace.ExportTraceServiceRequest{}, modelData)
+		return pdata.Traces{}, pdata.NewErrIncompatibleType(&otlpcollectortrace.ExportTraceServiceRequest{}, modelData)
 	}
 	return pdata.TracesFromInternalRep(internal.TracesFromOtlp(td)), nil
 }
